Extracción de entidades del modelo de Movie.kt a Entities.kt Extracción del método de clase RemoteMovie.toDomainMovie() a una función de extensión en EntitiesFunctions.kt Ajustes en los viewModels para que funcionen con los cambios anteriores
