+++ composeApp/src/desktopMain/kotlin/edu/dyds/movies/data/Entities.kt
@@ -1,0 +1,27 @@
+package edu.dyds.movies.data
+
+import edu.dyds.movies.domain.entity.Movie
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class RemoteResult(
+    val page: Int,
+    val results: List<RemoteMovie>,
+    @SerialName("total_pages") val totalPages: Int,
+    @SerialName("total_results") val totalResults: Int
+)
+
+@Serializable
+data class RemoteMovie(
+    val id: Int,
+    val title: String,
+    val overview: String,
+    @SerialName("release_date") val releaseDate: String,
+    @SerialName("poster_path") val posterPath: String,
+    @SerialName("backdrop_path") val backdropPath: String?,
+    @SerialName("original_title") val originalTitle: String,
+    @SerialName("original_language") val originalLanguage: String,
+    val popularity: Double,
+    @SerialName("vote_average") val voteAverage: Double,
+)
